+++ Makefile
@@ -1,7 +1,6 @@
-.PHONY: image tests lint_tests code_tests docs sdist testpypi pypi
+.PHONY : all tests docs lint_tests code_tests clean
 
-image : dev-requirements.txt
-	docker build -t pythonflow .
+all : tests docs
 
 tests : lint_tests code_tests
 
@@ -9,20 +8,11 @@ lint_tests :
 	pylint pythonflow
 
 code_tests :
-	py.test --cov pythonflow --cov-fail-under=100 --cov-report=term-missing --cov-report=html
+	py.test --cov pythonflow --cov-fail-under=100 --cov-report=term-missing --cov-report=html -v
 
 docs :
 	sphinx-build -b doctest docs build
 	sphinx-build -nWT docs build
-
-sdist :
-	python setup.py sdist
-
-testpypi : sdist
-	twine upload --repository-url https://test.pypi.org/legacy/ dist/pythonflow-*
-
-pypi : sdist
-	twine upload dist/pythonflow-*
 
 clean :
 	rm -rf build/
